fixes app fixutre
