+++ services/payments/tests/unit/conftest.py
@@ -6,12 +6,24 @@
 
 
 from collections.abc import AsyncIterator
+from typing import Callable
 
 import pytest
 from asgi_lifespan import LifespanManager
 from fastapi import FastAPI
+from pytest_mock import MockerFixture
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from simcore_service_payments.core.application import create_app
+
+
+@pytest.fixture
+def disable_rabbitmq_service(mocker: MockerFixture) -> Callable:
+    def _doit():
+        # The following moduls are affected if rabbitmq is not in place
+        mocker.patch("simcore_service_payments.core.application.setup_rabbitmq")
+        mocker.patch("simcore_service_payments.core.application.setup_rpc_api_routes")
+
+    return _doit
 
 
 @pytest.fixture
